Add orWhere() and andWhere() methods to Query class
This PR adds an orWhere() method to allow OR WHERE statements in queries which was not previously available. The PR also adds an andWhere() method which is an alias for where() (which would already produce an AND statement when chained on to another where() statement). The purpose of the andWhere() method is to optionally improve readability and clarity of fluent queries in the code.

The logic of the where() method has been refactored into a new method called addWhere() which when chained will add either an AND or an OR statement depending on the $operator argument which is passed in as the first argument. The where() method remains and behaves the same to keep the API consistent and backward compatible.

Queries can now be added like this.

    $this->xero
        ->load('Accounting\\Contact')
        ->where('Name', 'Steve')
        ->orWhere('Name', 'Nancy')
        ->execute();